simplify arguments and remove generate_utterances
- --generate_utterances is likely not going to be used. it makes things
much slower. removing it to simplify how arguments are handled
- remove --load_run and take the run from --load_checkpoint

Now there are two main ways to run rsa tests:
1. Skip training and load checkpoint from the given run folder (eg. '5').
If --test_rsa is 'train' or 'validation'. this will also load messages from the
interaction file in the same folder.
2. Run the training pipeline and use the freshly trained models. If
--test_rsa is 'train' or 'validation', it will save the interactions and then
load them during the rsa test phase.

In either scenario it runs the regular test first, and uses the interaction
returned from `.eval()` to get the messages for the RSA test run, if --test_rsa
is 'test'.