load listener & correct symbol informativeness
Added a function that loads the final listener from training. Also created a main: calc_metrics_lazy.ipynb that calculates and potentially saves the metrics for laziness. It is mostly a copy of evaluate_metrics.ipynb but there is an added part for ZLA significance and symbol information content. The latter should also be working correctly now, it had a few bugs, that I could recognize by wrong output when using trained data.